Use vagrant-vbguest for installing guest additions
VBoxGuestAdditions doesn't allow you to specify a kernel version which means
that they are always built against the pre-yum updated version and fail to
load when the guest is rebooted and exported.

Considered adding the updates repo to the KS to install within the point
release but there are various install-related bugs in doing so.